Build AMR vector

Build AMR from the result of function AMR_search

Usage

AMR_build(res, methylation, nb_cpg = 2)

Arguments

res

result object of function AMR_search

methylation

a matrix of methylation profile.

nb_cpg

threshold of minimal number of CpG in the AMR

Value

A set of build AMRs.

res, selected AMR

CpG_for_each_AMR, list of markers present on each AMR.

AMR_acp, first components of PCA for each DMR

AMR_mean, mean value of CpG on the AMR

Details

We use the series of pValues (one pValue per CpGs) obtained with the mEWAS regression method and the combination of pValue max2. To determine the potential AMRs used the combp method present in the ENmix package (Xu et al. 2016). This method uses the Fisher method to combine the pValues and also the base pair distance (bP) between CpGs (1000 bP maximum between nb_cpg CpGs on the same AMR). The information for each AMR is summarized by doing the mean (by row) of each CpG.

Author

Basile Jumentier

Examples

data = hdmax2::helper_ex
#Artificial reduction of dataset size to pass the github action check when building hdmax2 website
data$methylation = data$methylation[ , 800:1000]
data$annotation = data$annotation[800:1000, ]
K=5
## run hdmax2 step1
hdmax2_step1 = hdmax2::run_AS(X = data$exposure,
                             Y = data$phenotype,
                             M = data$methylation,
                             K = K)
#> [1] "The exposure argument is a vector."
#> [1] "The outcome argument is a vector."
#> [1] "The outcome argument is numeric"
#> [1] "Potential mediators matrix is actually a matrix"
#> [1] "provided K = 5"
#> [1] "K value has been transformed as integer"
#> [1] "The input exposome is continuous or binary"
#> [1] "The outcome vector is numeric and DON'T contains only 0s and 1s, it is assimilated as continous variable."
#> [1] "Running first regression with univariate explanatory variable."
#> [1] "Running second regression."
#> [1] "Running max-squared test."
#> [1] "Not generating max2 pvalues for each explanatory variable."

##Detecting AMR
chr = data$annotation$chr
start = data$annotation$start
end = data$annotation$end
pval = hdmax2_step1$max2_pvalues
cpg = data$annotation$cpg

res.amr_search = hdmax2::AMR_search(chr = data$annotation$chr,
                                   start = data$annotation$start,
                                   end = data$annotation$end,
                                   pval = hdmax2_step1$max2_pvalues,
                                   cpg = data$annotation$cpg,
                                   seed = 0.6, #Careful to change this parameter when working with real data
                                   nCores = 2)

res.amr_search$res
#>   chr   start     end           p        fdr
#> 2   1 1326166 1326221 0.005626678 0.01688003
#> 1   1 1310734 1310988 0.019799860 0.02138751
#> 3   1 1406862 1407016 0.021387509 0.02138751

res.arm_build = hdmax2::AMR_build(res.amr_search, 
methylation = data$methylation, nb_cpg = 2)
#List of DMR selected
head(res.arm_build$res)
#>    DMR chr   start     end          p        fdr nb
#> 1 DMR1   1 1310734 1310988 0.01979986 0.02138751  2
#> 2 DMR2   1 1406862 1407016 0.02138751 0.02138751  2
## CpG in the DMR
res.arm_build$CpG_for_each_AMR
#> $DMR1
#> [1] "cg11518257" "cg15617543"
#> 
#> $DMR2
#> [1] "cg10228629" "cg06377929"
#>