Identifying aggregated mediator regions (AMR)

Identify aggregated methylated regions (AMR) from the P-values from function max2 using a modified comb-p method. Compute the P-value and the FDR for each AMR detected.

Usage

AMR_search(chr, start, end, pval, cpg, ...)

Arguments

chr

chromosomes

start

chromosomal position of markers (start)

end

chromosomal position of markers (end)

pval

P-values for each markers, from the max2 function

cpg

name of each markers

...

see help of combp of ENmix package

Value

res, table of selected AMRs. For each AMR include chromosomic position, P-value, and FDR

data, matrix of all cpg, with annotation and provided P-values

Details

The function uses a modified comb-p method to identify aggregated methylated regions (AMRs).

Author

Basile Jumentier

Examples

data = hdmax2::helper_ex
#Artificial reduction of dataset size to pass the github action check when building hdmax2 website
data$methylation = data$methylation[ , 800:1000]
data$annotation = data$annotation[800:1000, ]
K=5
## run hdmax2 step1
hdmax2_step1 = hdmax2::run_AS(exposure = data$exposure,
                             outcome = data$phenotype,
                             M = data$methylation,
                             K = K)
#> Running first regression with univariate exposure variable.
#> Running second regression.
#> Running max-squared test.

##Detecting AMR
chr = data$annotation$chr
start = data$annotation$start
end = data$annotation$end
pval = hdmax2_step1$max2_pvalues
cpg = data$annotation$cpg

res.amr_search = hdmax2::AMR_search(
chr = data$annotation$chr,
start = data$annotation$start,
end = data$annotation$end,
pval = hdmax2_step1$max2_pvalues,
cpg = data$annotation$cpg,
seed = 0.6, #Careful to change this parameter when working with real data
nCores = 2)
res.amr_search$res
#>   chr   start     end           p        fdr
#> 2   1 1326166 1326221 0.005626678 0.01688003
#> 1   1 1310734 1310988 0.019799860 0.02138751
#> 3   1 1406862 1407016 0.021387509 0.02138751