Association Study with both exposure and outcome

This function uses lfmm (latent factor mixed models) to estimate the effects of exposures and outcomes on a response matrix. applied to estimate the effects of exposure $X$ on a matrix $M$ of potential mediators, and the effect of each marker on outcome $Y$. It uses the covariables matrix $conf$ and $K$ latent factors. Test all possible markers to determine potential mediators in the exposure-outcome association. Then compute the squared maximum of two series of P-values with $max2$ test It computes the squared maximum of two series of P-values from the association studie. This rejects the null-hypothesis that either the effect of X on M, or the effect of M on Y is null.

Usage

run_AS(
  X,
  Y,
  M,
  K,
  covar = NULL,
  detailed = FALSE,
  diagnostic.plot = FALSE,
  genomic.control = TRUE,
  effect.sizes = FALSE
)

Arguments

X

Exposure. An explanatory variable data frame with n rows and d columns. Each column corresponds to a distinct explanatory variable (Exposure). Continuous and binary variables must be encoded in numeric format. categorical variables are factor objects. The user can use the as.factor function to encode categorical variables, and levels() and ordered() functions to define the modal order of categorical variables.

Y

Outcome. An explanatory variable matrix with n rows and 1 columns, corresponds to a vector, which supports both continuous and binary formats.

M

a response variable matrix with n rows and p columns. Response variables must be encoded as numeric. No NAs allowed.

K

an integer for the number of latent factors in the regression model.

covar

set of covariable, must be numeric. No NAs allowed

detailed

A logical to indicate if p-values must be estimated for each explanatory variables (detailed = TRUE) in addition to the pvalue of the global model (detailed = FALSE, by default)

diagnostic.plot

if TRUE the histogram of the p-values together with the estimate of eta0 null line is plotted. Useful to visually check the fit of the estimated proportion of null p-values.

genomic.control

correct pvalue with genomic inflation factor

effect.sizes

if effect sizes from lfmm are needed

Value

an object with the following attributes

for first association study (mod1):

pValue, estimation of the effects of X and Y on the matrix M.

U, scores matrix for the K latent factors computed from the for first regression

zscore, a score matrix for the exposure X and the outcome Y.

fscore, a score matrix for the exposure X and the outcome Y.

adj_rsquared

gif, Genomic inflation factor for X and Y, expressing the deviation of the distribution of the observed test statistic compared to the distribution of the expected test statistic

for second association study (mod2):

pValue, zscore, fscore, adj_rsquared, gif

results of max2 test:

pval, results of max2 test

input element: exposition , outcome, matrix (element and type) and covar

Details

The response variable matrix Y and the explanatory variable are centered. For each argument, missing values must be imputed: no NA allowed. K (number of latent factors) can be estimated with the eigenvalues of a PCA. Possibility of calibrating the scores and pValues by the GIF (Genomic Inflation Factor). See LEA package for more information. Max2 test The P-value is computed for each markers following this formula $$pV = max(pVal1, pVal2)^2$$

Author

Florence Pittion, Magali Richard

Examples

# Load example dataset
simu_data = hdmax2::simu_data

# Run {hdmax2} step 1
hdmax2_step1 = run_AS(X = simu_data$X_continuous ,
                      Y =  simu_data$Y_continuous,
                      M =  simu_data$M,
                      K=5)
#> [1] "The exposure argument is a vector."
#> [1] "The outcome argument is a vector."
#> [1] "The outcome argument is numeric"
#> [1] "Potential mediators matrix is actually a matrix"
#> [1] "provided K = 5"
#> [1] "K value has been transformed as integer"
#> [1] "The input exposome is continuous or binary"
#> [1] "The outcome vector is numeric and DON'T contains only 0s and 1s, it is assimilated as continous variable."
#> [1] "Running first regression with univariate explanatory variable."
#> [1] "Running second regression."
#> [1] "Running max-squared test."
#> [1] "Not generating max2 pvalues for each explanatory variable."

head(hdmax2_step1$max2_pvalues)
#> cg00005740 cg00006787 cg00007032 cg00008612 cg00009871 cg00009922 
#>  0.2614245  0.4061289  0.1119728  0.9231155  0.7612041  0.8804220